Start migrating configuration to a new structure
This is the first step in using a different configuration strategy for
tracking the user's input.

One goal is to use meaningful identifiers for configuration data to make
the code easier to read.

Another goal is to be able to split the code into different modules to
make it easier to maintain, and this is a preliminary step toward that.