Try a different approach for pynput issue
Philippe on Stack Overflow suggests at
https://stackoverflow.com/a/77294321/104848 that the issue might be
that the keyboard.Listener gets started in the button's call.

The approach he suggests is to start the Listener before running any of
the Tk code. This works, but it means that the app will always be
listening to and handling keystrokes.

As an alternative, this test script runs the pynput listener in a
separate thread. However, this script crashes python in the same way.